- AppleClang error: SIMDPP error; seems to be solved with dev (check on compilation) - windows error: see if wrapping in numpy array resolves the issue

+++ pybinding/lattice.py
@@ -176,7 +176,7 @@ class Lattice:
             self.add_one_alias(name, alias, position)
         else:
             # convert single-element arrays to scalars
-            energy = np.asarray(onsite_energy)
+            energy = np.asarray(onsite_energy, dtype=float)
             if energy.size == 1:
                 energy = energy.item()
             self.impl.add_sublattice(name, position, energy)
